R.U.B.E.-Level triggered collision Sound effects. Fixtures and bodies now support the customPropertie collisionSound which holds the name of an mp3 effect file. The Classes ChipmunkSegmentShape, ChipmunkPolyShape and ChipmunkCircleShape have subclasses supporting the CMObjectSoundProtocol (simple string propertie)

+++ MarbleGame/Level/RUBEReader/CMRubeSceneReader.m
@@ -52,7 +52,7 @@
 	// now attach the images to the associated Bodies
 	for (CMRubeImage *rImage in self.images) {
     if (rImage.rubeBodyIndex != NSNotFound) {
-			NSLog(@"Image %@ to %ld",rImage.name,rImage.rubeBodyIndex);
+			NSLog(@"Image %@ to %ld",rImage.name,(long)rImage.rubeBodyIndex);
 			CMRubeBody* aBody = [self.bodies objectAtIndex:rImage.rubeBodyIndex];
 			[aBody attachImage:rImage];
 		}
